add: lightとcameraの切り出し,fix: キー入力を直感的に,remove: マウス入力によるカメラの移動の無効化
