+++ app/page.tsx
@@ -1,6 +1,5 @@
 'use client';
 
-import { OrbitControls, PerspectiveCamera } from '@react-three/drei';
 import { Canvas } from '@react-three/fiber';
 import { EffectComposer, SSAO } from '@react-three/postprocessing';
 import { XR, ARButton } from '@react-three/xr';
@@ -8,6 +7,8 @@ import React, { useState } from 'react';
 import { Vector3 } from 'three';
 import { PlateauTileset } from './_components/PlateauTileset';
 import { PlateauTilesetTransform } from './_components/PlateauTilesetTransform';
+import { ThreeCamera } from './_components/ThreeCamera';
+import { ThreeLight } from './_components/ThreeLight';
 
 const Home: React.FC = () => {
   const [cameraPosition, setCameraPosition] = useState<Vector3>(new Vector3(1050, 12, -250));
@@ -17,17 +18,10 @@ const Home: React.FC = () => {
       <ARButton />
       <Canvas>
         <XR>
-          <PerspectiveCamera makeDefault position={cameraPosition} near={1} far={1000} />
-          <OrbitControls target={cameraPosition} />
-          <directionalLight
-            position={[cameraPosition.x, cameraPosition.y + 100, cameraPosition.z]}
-            intensity={1}
-            castShadow
-            shadow-mapSize={[8192, 8192]}
-          >
-            <orthographicCamera attach='shadow-camera' args={[-2500, 2500, 2500, -2500, 1, 5000]} />
-          </directionalLight>
-          <PlateauTilesetTransform setCameraPosition={setCameraPosition}>
+          <fogExp2 attach='fog' color='white' density={0.0002} />
+          <ThreeCamera cameraPosition={cameraPosition} setCameraPosition={setCameraPosition} />
+          <ThreeLight cameraPosition={cameraPosition} />
+          <PlateauTilesetTransform>
             {/* <PlateauTileset path='bldg/23100_nagoya/23101_chikusa-ku/notexture' /> */}
             <PlateauTileset path='bldg/23100_nagoya/23102_higashi-ku/notexture' center />
           </PlateauTilesetTransform>
